Update README.md
This project studies the data analysis of assets and structures under the responsibility of the mayor of the 5th district of Douala such as electricity poles, schools, hospitals, street lights, traffic lights, gutters, gaming and money transfer kiosks, etc. ...... Here we study each asset separately as well as the location from which it was collected and the condition in which it is found. 

This work will allow the city council to know exactly where the structures and objects in its charge are located and in what condition they are in order to assess whether they are harmful to the inhabitants or not.